Cleanup of auth module.
* Organise imports and formatting.
* Log uncaught exceptions in data layer.
* Typing cleanup and explicit assertions.
* Move auth to module.
* Factor out JWT functionality.
